Refactor Form with Navigation
