+++ src/modules/admin/components/forms/FormPreview.tsx
@@ -1,4 +1,4 @@
-import { Button, Grid } from '@mui/material';
+import { Button } from '@mui/material';
 import { Stack } from '@mui/system';
 import { useCallback, useMemo, useRef, useState } from 'react';
 import { generatePath } from 'react-router-dom';
@@ -19,16 +19,16 @@ import SentryErrorBoundary from '@/modules/errors/components/SentryErrorBoundary
 import DynamicForm, {
   DynamicFormRef,
 } from '@/modules/form/components/DynamicForm';
-import FormNavigation from '@/modules/form/components/FormNavigation';
+import FormNavigationContainer from '@/modules/form/components/FormNavigationContainer';
 import DynamicView from '@/modules/form/components/viewable/DynamicView';
 import { FormValues } from '@/modules/form/types';
 import {
   AlwaysPresentLocalConstants,
   createInitialValuesFromSavedValues,
+  createValuesForSubmit,
   getFormStepperItems,
   getInitialValues,
   getItemMap,
-  createValuesForSubmit,
 } from '@/modules/form/util/formUtil';
 import { RootPermissionsFilter } from '@/modules/permissions/PermissionsFilters';
 import { AdminDashboardRoutes } from '@/routes/routes';
@@ -207,13 +207,9 @@ const FormPreview = () => {
       </Stack>
 
       <SentryErrorBoundary>
-        {formStepperItems ? (
-          <Grid container spacing={2} sx={{ pb: 20, mt: 0 }}>
-            <FormNavigation items={formStepperItems}>{form}</FormNavigation>
-          </Grid>
-        ) : (
-          form
-        )}
+        <FormNavigationContainer navItems={formStepperItems}>
+          {form}
+        </FormNavigationContainer>
       </SentryErrorBoundary>
     </>
   );
