+++ src/modules/client/components/pages/EditClientPage.tsx
@@ -1,11 +1,6 @@
-import { Box, Stack, Typography } from '@mui/material';
+import { Stack, Typography } from '@mui/material';
 import { useCallback, useMemo } from 'react';
 import { useNavigate } from 'react-router-dom';
-
-import {
-  CONTEXT_HEADER_HEIGHT,
-  STICKY_BAR_HEIGHT,
-} from '@/components/layout/layoutConstants';
 import DeleteClientButton from '@/modules/client/components/DeleteClientButton';
 import useClientDashboardContext from '@/modules/client/hooks/useClientDashboardContext';
 
@@ -42,7 +37,6 @@ const EditClientPage = () => {
       record={client}
       localConstants={localConstants}
       onCompleted={onCompleted}
-      top={STICKY_BAR_HEIGHT + CONTEXT_HEADER_HEIGHT}
       title={
         <Stack direction='row' justifyContent='space-between'>
           <Typography
@@ -57,17 +51,13 @@ const EditClientPage = () => {
       }
       FormNavigationProps={{
         contentsBelowNavigation: (
-          <Box sx={{ mt: 3 }}>
-            <RootPermissionsFilter permissions='canDeleteClients'>
-              <DeleteClientButton
-                clientId={client.id}
-                clientLockVersion={client.lockVersion}
-                onSuccess={() =>
-                  navigate(generateSafePath(Routes.CLIENT_SEARCH))
-                }
-              />
-            </RootPermissionsFilter>
-          </Box>
+          <RootPermissionsFilter permissions='canDeleteClients'>
+            <DeleteClientButton
+              clientId={client.id}
+              clientLockVersion={client.lockVersion}
+              onSuccess={() => navigate(generateSafePath(Routes.CLIENT_SEARCH))}
+            />
+          </RootPermissionsFilter>
         ),
       }}
     />
